Travis: downgrade scipy
Travis builds started failing with "libstdc++.so.6: version
`GLIBCXX_3.4.21' not found" when scipy 0.17.0 moved from np110pyxx_2 to
np110pyxx_3. This added libgcc-5.2.0 to the conda dependencies, which
may be the cause of the trouble (only other chance was libfortran-3.0 to
libfortran 3.0.0, build 1).

I have not been able to reproduce this problem on my system (which
however runs ubuntu 14 rather than 12).