Add base saturation region reference treshold (#566)
* Add min number of reference samples before desat

This makes sure that we only desaturate a channel if we have enough reference samples above 1PE to work with. This distinguishes between S2, which will still be desaturated, and S1, which do not get desturated, signals.

* Adds the value: reference_region_samples_treshold

Here we set the minimal number of reference samples above 1PE to 10. More info is in the pull request.

* whitespaces

* more white spaces