Don't keep LED data in buffer (#600)
The LED trigger config is hardcoded in the eventbuilder, not specified by the run doc. After the changes to how we run the eventbuilder following #568, LED data is no longer automatically deleted, but only after some time (just like background data). Obviously we're not going to detect supernovae in LED calibration mode, so let's not do this.