More reasonable electron lifetime default (#647)
The electron lifetime default was from XENON100.  This is overloaded at the rundb and in hax.  In case people run pax on its own, we should have a sensible default so they don't go crazy looking at the cs2.  This is related to changes in https://github.com/XENON1T/cax/pull/143